+++ cli/src/commands/workspace/add.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::fmt::Display;
 use std::fs;
 
 use itertools::Itertools;
@@ -31,9 +32,31 @@ use crate::command_error::user_error;
 use crate::command_error::CommandError;
 use crate::ui::Ui;
 
+/// How to handle sparse patterns when creating a new workspace.
+#[derive(clap::ValueEnum, Clone, Debug, Eq, PartialEq)]
+enum SparseInheritance {
+    /// Copy all sparse patterns from the current workspace.
+    Copy,
+    /// Include all files in the new workspace.
+    Full,
+    /// Clear all files from the workspace (it will be empty).
+    Empty,
+}
+
+impl Display for SparseInheritance {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            SparseInheritance::Copy => write!(f, "copy"),
+            SparseInheritance::Full => write!(f, "full"),
+            SparseInheritance::Empty => write!(f, "empty"),
+        }
+    }
+}
+
 /// Add a workspace
 ///
-/// Sparse patterns will be copied over from the current workspace.
+/// By default, the new workspace inherits the sparse patterns of the current
+/// workspace. You can override this with the `--sparse-patterns` option.
 #[derive(clap::Args, Clone, Debug)]
 pub struct WorkspaceAddArgs {
     /// Where to create the new workspace
@@ -58,6 +81,9 @@ pub struct WorkspaceAddArgs {
     /// new r1 r2 r3 ...`.
     #[arg(long, short)]
     revision: Vec<RevisionArg>,
+    /// How to handle sparse patterns when creating a new workspace.
+    #[arg(long, default_value_t=SparseInheritance::Copy)]
+    sparse_patterns: SparseInheritance,
 }
 
 #[instrument(skip_all)]
@@ -114,19 +140,29 @@ pub fn cmd_workspace_add(
         )?;
     }
 
-    // Copy sparse patterns from workspace where the command was run
     let mut new_workspace_command = command.for_workable_repo(ui, new_workspace, repo)?;
-    let (mut locked_ws, _wc_commit) = new_workspace_command.start_working_copy_mutation()?;
-    let sparse_patterns = old_workspace_command
-        .working_copy()
-        .sparse_patterns()?
-        .to_vec();
-    locked_ws
-        .locked_wc()
-        .set_sparse_patterns(sparse_patterns)
-        .map_err(|err| internal_error_with_message("Failed to set sparse patterns", err))?;
-    let operation_id = locked_ws.locked_wc().old_operation_id().clone();
-    locked_ws.finish(operation_id)?;
+
+    let sparsity = match args.sparse_patterns {
+        SparseInheritance::Full => None,
+        SparseInheritance::Empty => Some(vec![]),
+        SparseInheritance::Copy => {
+            let sparse_patterns = old_workspace_command
+                .working_copy()
+                .sparse_patterns()?
+                .to_vec();
+            Some(sparse_patterns)
+        }
+    };
+
+    if let Some(sparse_patterns) = sparsity {
+        let (mut locked_ws, _wc_commit) = new_workspace_command.start_working_copy_mutation()?;
+        locked_ws
+            .locked_wc()
+            .set_sparse_patterns(sparse_patterns)
+            .map_err(|err| internal_error_with_message("Failed to set sparse patterns", err))?;
+        let operation_id = locked_ws.locked_wc().old_operation_id().clone();
+        locked_ws.finish(operation_id)?;
+    }
 
     let mut tx = new_workspace_command.start_transaction();